🔵 Move discount inside product domain
