--- src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/discount/rules/DiscountRule.kt
@@ -1,8 +1,0 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount.rules
-
-import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountPercentage
-import com.capitole.capitoleproductcatalogapi.domain.product.Product
-
-interface DiscountRule {
-  fun applyTo(product: Product): DiscountPercentage
-}
