+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/DefaultDiscountService.kt
@@ -1,7 +1,7 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount
+package com.capitole.capitoleproductcatalogapi.domain.product.discount
 
-import com.capitole.capitoleproductcatalogapi.domain.discount.rules.DiscountRule
 import com.capitole.capitoleproductcatalogapi.domain.product.Product
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.rules.DiscountRule
 
 class DefaultDiscountService(private val rules: List<DiscountRule>) : DiscountService {
   override fun getApplicableDiscount(product: Product): DiscountPercentage =
