+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/DiscountPercentage.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount
+package com.capitole.capitoleproductcatalogapi.domain.product.discount
 
 data class DiscountPercentage(val value: Double) {
 /*
