+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/DiscountService.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount
+package com.capitole.capitoleproductcatalogapi.domain.product.discount
 
 import com.capitole.capitoleproductcatalogapi.domain.product.Product
