+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/rules/CategoryRule.kt
@@ -1,8 +1,8 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount.rules
+package com.capitole.capitoleproductcatalogapi.domain.product.discount.rules
 
-import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountPercentage
 import com.capitole.capitoleproductcatalogapi.domain.product.Category
 import com.capitole.capitoleproductcatalogapi.domain.product.Product
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.DiscountPercentage
 
 class CategoryRule(private val target: Category, private val discountPercentage: Double) : DiscountRule {
   override fun applyTo(product: Product) =
