+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/rules/DiscountRule.kt
@@ -1,0 +1,8 @@
+package com.capitole.capitoleproductcatalogapi.domain.product.discount.rules
+
+import com.capitole.capitoleproductcatalogapi.domain.product.Product
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.DiscountPercentage
+
+interface DiscountRule {
+  fun applyTo(product: Product): DiscountPercentage
+}
