+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/discount/rules/SkuEndsWith5Rule.kt
@@ -1,7 +1,7 @@
-package com.capitole.capitoleproductcatalogapi.domain.discount.rules
+package com.capitole.capitoleproductcatalogapi.domain.product.discount.rules
 
-import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountPercentage
 import com.capitole.capitoleproductcatalogapi.domain.product.Product
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.DiscountPercentage
 
 class SkuEndsWith5Rule : DiscountRule {
   override fun applyTo(product: Product) =
