+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/configuration/ApplicationConfiguration.kt
@@ -1,8 +1,8 @@
 package com.capitole.capitoleproductcatalogapi.infrastructure.configuration
 
 import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.dto.GetProductCatalog
-import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountService
 import com.capitole.capitoleproductcatalogapi.domain.product.ProductRepository
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.DiscountService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
