+++ src/test/kotlin/com/capitole/capitoleproductcatalogapi/domain/discount/rules/CategoryRuleTest.kt
@@ -1,11 +1,12 @@
 package com.capitole.capitoleproductcatalogapi.domain.discount.rules
 
-import com.capitole.capitoleproductcatalogapi.domain.discount.DiscountPercentage
 import com.capitole.capitoleproductcatalogapi.domain.product.Category
 import com.capitole.capitoleproductcatalogapi.domain.product.Description
 import com.capitole.capitoleproductcatalogapi.domain.product.Price
 import com.capitole.capitoleproductcatalogapi.domain.product.Product
 import com.capitole.capitoleproductcatalogapi.domain.product.SKU
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.DiscountPercentage
+import com.capitole.capitoleproductcatalogapi.domain.product.discount.rules.CategoryRule
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
